Config updates relating to DIB-695
Changed the majority of the configurations (that we own) in the configuration tab of the
web console to follow a standard naming convention. When adding new
configurations, the name in its metatype file should start out with the general area the
configuration belongs too (catalog, content, security, etc..) and then
follow with the name of the thing it is configuring, i.e. Catalog File
System Storage Provider.