DDF-241 Catalog Frameworks fixed to use new Catalog implementations
Standard and Fanout frameworks were changed and their tests fixed.
Moved the CatalogFrameworkImpl into the standard catalog framework project.
FanoutCatalogFramework moved into the fanout framework project.
Tests moved to their respective projects. The fanout project depends on the
standard framework project. FanoutCatalogFramework package was changed to not conflict with
the standard framework when it came to importing and embedding packages.