DDF-373 Updated test data to reflect latest application service response, fixed IDs with periods, handled combined app name and version.
Replaced test data to mirror latest response from the application service.
Handled application names with version numbers included - updated version to display in details correctly.
Added an appId to the model and set it on initialization, cleaned up the display name and version in the special case where they are combined.
Updated test to check the combined app name and version case.