DDF-311 Add extra temporal filters
This is a reworked version based on feedback from Phil Klinefelter. All issues except for subclassing
GeotoolsFilterBuilder have been addressed.

Tests originally checked the boundary condition (the equal-to case), however it was only about 80% reliable,
and adding a small offset defeated the value of the tests, so only simple before and after checks are included.