DDF-705 - Converting Configuration Module to System Module.
Converting the Configuration Module to System. As there are changes planned for the modules down the road, the only change is visually for now.
Added the same 'plugin' framework that applications have to modules. Added System Info and Configuration page to the system module.
System Info has some information about the system.
Configuration is what was previously shown by the Configuration module.