DDF-1319: Fixed BasicAuth support for OpenSearch.
BasicAuth was broken during the initial implementation of the
SecureCxfClientFactory in OpenSearchSource. It was trying to get a secure
client only using the security subject. It now properly uses the username and
password to get a secure client using basic auth and then attempts to apply the
security subject on that client.