DDF-1426 - Fixed Directory Monitor Security error
Recent changes that check for a security subject on a thread broke the
Directory Monitors ingest (Since it has no security subject) and was
throwing an uncaught runtime exception. Now we catch the exception to
allow the ingest to proceed and simply don't add a point of contact to
the metacard.