DDF-1448 Refactors most of functionality out of AbstractIntegrationTest for greater reusability
Most of the functionality of the base class has been pulled out to the new facade classes. In order for it to be initialized, PaxExamRule has been modified slightly to run a post-construction method. The methods that are now delegating to the new facade classes have all been deprecated with the intent that we will reroute all those calls directly to the appropriate class.