DDF-1493 Remove gazetteer feature from search-ui-app feature.
Prior to this change, the webservice-gazetteer was being loaded as part of the
search-ui feature.  While the search-ui needs a gazetteer, the selection of a
specific gazetteer implementation should be left to configurations that are
loaded at a later time.  This change moves the gazetteer loading into the
install-profiles/*/features.xml.