DDF-1539: Radius value is set correctly after a search
The “radiusValue” input field was receiving a default Backbone binding
because default bindings are assigned to elements with a name
attribute, and the input field had one. This binding overwrote the
value for the input field given by the binding for the hidden “radius”
field, which is the binding we want to control the view. Removing the
“radiusValue” input field’s name attribute prevents this default
binding from being instantiated.
This change required us to update the existing code that interacts with
the input field element to select it by id rather than by name.