DDF-1504: Fix GetRecordById response creation
-The GetRecordById request does not have a ResultType, so the CSW query
response transformation will not fail if the request is by ID and does
not have a ResultType.
-Previously, CswQueryResponseTransformer was always setting the root
node of the response to GetRecordsResponse even if the request was by
ID, in which case it should be GetRecordByIdResponse. Now, the root
node of the response is GetRecordByIdResponse if the request is by ID.