DDF-1570 Latitude/longitude ordering should be handled more consistently for CSW records
The CSW endpoint assumes incoming bounding box values are in lon/lat order and reverses
them before ingesting. The endpoint should use lat/lon order exclusively, only changing
the order when the data comes from a federated system which has been configured for lon/lat.

This change fixes a bug in the BoundingBoxReader class which was reversing the axis order
when it shouldn't be and not reversing it when it should.  Also updated tests.