DDF-1519 Encode relay state before sending to IdP
Relay state cannot exceed 80 characters and it should not leak
information about the user action.  Relay state is now a GUID that
can be used to lookup the original redirect url after the user has
successfully logged into the IdP.  Relay state will expire if the
user does not return.