DDF-1571 Adds integration tests for persisting to workspace.
An issue has been found during a demo in a production environment where attempting to save a large workspace fails with an error. This is caused by Lucene, which changed its behavior from logging and ignoring a too-large field to throwing an error if a field of >32776 bytes is indexed. We should not be indexing the data we put in the persistent store, but are for now. Two tests are added here: one with a small data size and one too large; the latter has been marked Ignored for now.