DDF-1572 Shuts off periodic bundlefile backups for itest exam folders.
We see an intermittent failure with itests where the entire pax exam setup fails, causing all tests in a class to fail. Errors have been seen in the osgi code when it attempts to create periodic backups of the `.bundlefile` in `${ddf.home}/data/cache/org.eclipse.osgi`.

Working on the supposition that pax can explode the exam directly successfully but that on slow boxes the background backup thread may start too soon, this will turn off those periodic backups. They are not necessary for itests; this will have either a positive or neutral effect.

It is possible that this is only a symptom of a more basic problem exploding the exam folder in which case this will merely shift the error.