DDF-1543 Fix various bugs
-Switching from USNG to Lat Lon would not properly update the Lat Lon values.
The template now looks to the 'map' version of these variables as opposed to
the raw versions. This continues the migration from 'north' -> 'mapNorth'
that was started in these USNG changes. The raw variables just track where the
cursor is and does not decide what is shown on the map or what is searched.

-Updated search function to use the 'map' versions of variables. Silently sets
the raw variables to the 'map' variables to not trigger a redraw but to affect
other code as little as possible - best not to percolate the 'map' change, except
where needed.