DDF-1660 Fix ConfigurationEdit to set Boolean
Previously ConfigurationEdit looked for the existence of a variable
bindValue and then would set bindValueString based on that. Now if
bindValue is "falsey" bindValueString will be set to "false" instead of
"undefined".

DDF-1660 Revert metatype

DDF-1660 Refactor code to make smoother

DDF-1660 Make service force boolean to false if null

If someone forgets to set a default boolean value in the metatype, it will
be set to null. The javascript backbone model handles this fine, but when
sent back to the backend, it causes failures. Now, if a model gets has no
default value and should be a boolean, it is set to false

DDF-1660 Refactor logic to be clearer