DDF-1660 Make service force boolean to false if null
If someone forgets to set a default boolean value in the metatype, it will
be set to null. The javascript backbone model handles this fine, but when
sent back to the backend, it causes failures. Now, if a model gets has no
default value and should be a boolean, it is set to false