DDF-1669: Update owasp to run on maven validate phase
Since there is no way to omit just the distribution/test
directory, and not all dependencies can be set to a test scope,
owasp has to piggy-back on a maven phase. Since validate takes
a trivial amount of time, using owasp with this phase as
"mvn validate -Powasp" will essentially be synonymos with running
owasp as an individual goal, while allowing modules to avoid running
the scan by specifying the phase "none" for the owasp plugin.