DDF-1875 Fix Jacoco floating point comparison issue with platform-scheduler line coverage
When building locally, Jacoco passes the platform-scheduler module. When building on Bamboo, it fails with the error
```
lines covered ratio is 0.35, but expected minimum is 0.36
```
In testing this out, I changed the expected minimum locally to 0.37 and received the error
```
lines covered ratio is 0.35, but expected minimum is 0.37
```
Clearly on my OS (and others') the floating point comparison between 0.35xxxxx and 0.36xxxxx is coming up true whereas on the Bamboo server it is returning false. Since the coverage number is being reported by Jacoco as 0.35, it should be dropped to match reality.