DDF-1648 Update the manner in which the ConfigurationAdmin properties are updated at initial boot.
Instead of updating the properties during a property set, attaches a service listener on initial boot waiting for the `BootFinished` service. When that service is published by Karaf, the boot sequence has completed and the config properties can safely be updated. On subsequent boots of the system, the listener is not attached.

Additionally, this cleans up to unrelated pom issues where `Private-Package` instructions where including packages that did not need to be imported to the bundles.