DDF-1974 Created catalog-transformer-pptx for powerpoint documents
The catalog-transformer-pptx feature uses the existing Tika input
transformer to generate the basic metadata, and then uses Apache
POI to generate a thumbnail image based on the first page of the
powerpoint deck. The documentation was updated to include the
catalog-transformer-pptx feature. The transformer only works for
the newer style OOXML Powerpoint (PPTX) files. It does not handle
the older style OLE2 Powerpoint (PPT) files.