DDF-2027: Small Logviewer Improvements
 - Reordered log level filters to match karaf levels
 - Added a link to open the viewer in a new tab
 - Renamed export button to be less ambiguous
 - Updated Export documentation and screenshots

  We did not move the export button under the configuration
tab because the code used to generate that tab is reused
in many other pages in the ui, and allowing for this specific
tab to display an export button would result in a massive amount
of code rewriting for very little value. Since the problem we were
trying to solve was the button's ambiguity, we decided that
the best thing to do was to relabel it to say "Export System
Configurations" to solve that issue.