DDF-2071 Update CatalogFrameworkImpl so callers can supply a fully populated metacard and bypass the input transformers.
If the ContentItem being stored contains a non-null metacard, then that metacard will be used instead of using
the input transformers to create a metacard. The non-test callers were already passing a null metacard. A few
test classes were supplying a non-null metacard, but those values were not actually used within the framework,
so they were changed to null.