ddf-2095 ddf should prefer /dev/urandom on Linux
This change adds logic to the setenv script to select the /dev/urandom device
on Linux hosts. This device is used to generate a seed value for encryption modules.
The JVM uses /dev/random by default, which can hang for very
long periods and cause a timeout in the web apps.