DDF-2188 Adds new service definition for extracting metadata from text content files.
The Tika and PDF input transformers have been updated to find the `ContentMetadataExtractor` implementors and process each of them in turn.

This change also updated the `MetacardAttributeSecurityPolicyPlugin` so that it takes key-value pairs of attribute names. The metacard is checked to see if it has attributes named by those keys and then creates security policy map entries keyed by the values.