DDF-2244 Add batching to the validation command and allow it to handle (#916)
* DDF-2244 created validate command API components

* DDF-2244 Stripped validateCommand and moved execution and printing to new classes

* DDF-2244 Validate command basically working, just need to fill out printing

* DDF-2244 command functioning properly. Need to redo unit tests

* DDF-2244: Added unit tests for the ValidateCommand

Added unit tests to test the functionality of ValidateCommand

Added command line options --path and --query to specify whether the user
will provide a file or directory to be validated, or if the user will pass
a query for the catalog to be searched for metacards to validate.

Validate command can now successfully use queried metacards for
validation.

To verify, test using the catalog:validate command to perform validation
on the results

Added support for CQL queries instead of text based queries, added support for
switching recursion subdirectories on or off, added support for choosing which
file extensions to search for in directories, and added paralellism support in
ValidateExecutor. ValidateExecuter.execute now allows parallelism exceptions
to be handled on a higher level, and includes the exceptions in the method
header. Simplified if statements in assigning value to files variable in
ValidateCommand.getFiles.