DDF-2391: Log Viewer Filtering Includes Higher Severities
Modified the log level filtering on the log viewer to reflect
how karaf filters logs - anything logs with the selected filtering
level or higher will be displayed.

Also fixed a formatting bug introduced with DDF-2270. That ticket
was supposed to fix UIs that went out to google to retrieve fonts,
however the log viewer does not load any additional fonts, and the
changes made broke the LogViewer's formatting, so the changes in
the logviewer were reverted.