DDF-2422: TestFederation iTest Intermittent Failures
Multiple tests inside TestFederation are failing intermittently.
One was fixed by setting the polling interval of the cswStubServer
source to 10min to match it up with the other sources. Further
investigation should be done to figure out why DDF marks sources
as unavailable when they are polled in iTests. This issue also
occurs with the other sources, and is most likely why they set
their polling intervals for that long.

The second test was fixed by moving the cometD message collection
call inside the expect().while() check. It had been left outside of
the loop which caused no messages to be recieved if they hadn't been
collected in the first call.