DDF-2453: LogViewer Pause/Resume Polling
After dicussing the use-case for this situation, we decided
that the best solution would be to add a pause/play button
to allow the admin to pause log polling in order to prevent
both the window from scrolling, as well as ensure that the
finite logviewer queue does not eventually push the logs
under inspection out of the finite queue.

Also cleaned up the redux code by using redux-thunk middleware
and combineReducers() to make future additions easier.