DDF-2297 Federated queries sometimes fail to return results (#1215)
This change fixes 2 bugs in the processing of federated search results.
First, it fixes a threading bug in SearchController that was blocking results
from being returned to the UI. Code had been added to handle exceptions, but
that code wasn't running in its own thread. As a result, the error-checking
prevented the method from returning.

Second, it fixes a minor bug in the case when the local solr results are
empty and those results weren't being merged into the set of cached
results. This was preventing the progress bar from being updated correctly.
(cherry picked from commit 362e9e0)