DDF-2468 Add support for more FTP commands (#1260)
STOU - Full support is added. The FTP server will return a success message that
includes the metacard id, which the client can interpret as the unique name given
to the object.

SITE - Since the FTP server does not have any site-specific commands, the server
now returns the status code 202, which indicates that SITE is not implemented.

DELE, RETR, RMD, MKD, APPE, RNTO - The FTP server returns a response indicating
success, even though no action was taken. Some clients, for example, always
attempt to make a new directory and then CD to it. This isn't necessary from
the DDF point of view. Previously, the server would return an error that would
cause the client to fail.

Add support for clients that PUT a dot-file and then rename the dot-file to
the final filename. The rename command will result in the object being stored
in the catalog.

Add itest for clients that the create a directory, change to that directory, and
then PUT content.

The CWD always returns a success response to support clients that use a
MKDIR/CWD/PUT command sequence.