DDF-2515 Deduplicates command argument alias. (#1357)
During a refactor and extension of functionality, the `DumpCommand` was made to extend a new `CqlCommands` base class, giving it additional query flexibility. That change duplicated the `-t` alias for a command argument. This changes the alias for the `--temporal` argument to be `-dt` so it will not conflict with the `--transformer` alias.