DDF-3167 Changes the timing for the IdP server & client initialization of metadata. (#2149)
Previously, these operations occurred on injection of the metadata locators. That led to an intermittent timing issue, as the two bundles cross-query HTTP endpoints. Now, they execute the queries and parse the data once, on initial access. This allows both services to be fully initialized before any queries are executed.

This change also includes some minor cleanup to MetadataConfigurationParser class. That cleanup had been associated with an earlier approach to resolving the timing issue and is not necessary; however, it does make for smaller, more concise methods.