DDF-3034 Prevent duplicate results during result pagination and handle filtered results.
Because we simulate a cursor by executing sequential queries with an incremented start index position, it is possible that ingests can occur, interleaved within subsequent queries. This could result in duplicate results appearing in multiple pages of the results. This will remove those duplicates.

This will also correctly page when there are filtered results, handling the case where a single page's worth of results is blanked out because all N results are filtered.