DDF-3484 HitCount-only Queries
Modified AtomTransformer to remove any returned entries when the 'count' parameter on the request is 0.
Modified QueryOperations and SolrMetacarClientImpl to pass along 'count=0' to CatalogProvider. Because
there may be installed systems relying on 'count=0' to return all results instead of hitcounts-only,
added a system property to make the default behavior configurable.