DDF-3105 Create Component test for REST endpoint
Changed how the PortFinder is being passed around in the AbstractComponentTest class.
Renamed the catalog core API and mime type API features to avoid conflict with those defined in the platform-app and catalog-app features files, which was causing tests to fail in some circumstances.
Improved some of the comments and removed unneeded dependencies in pom files.