DDF-3664 Made ddf-theme.css web accessible and pointed old bootswatch links to it.
The ddf-theme module now outputs a ddf-theme.css file with the Google font removed
and it is now web accessible. I started work on moving all of the old links to the
bootswatch css files to point to the new ddf-theme.css. Unfortunately, the pathing
for some of these may need adjustment. Currently the ddf-theme is exposed under the
search folder but the pattern in the code appears to have the old css served from
multiple places.