DDF-3664 Added bootswatch cosmo and cyborg themes to ddf-theme and added ddf-theme to login page.
In most of the application the bootswatch flatly theme is used, but in some places the cosmo and cyborg themes are used. I added the cosmo and cyborg themes to ddf theme and then linked the modules that used them to it using the link for the ddf.ui.search.simple module. Unfortunately, that is inaccessible on the login page. While the ddf-theme module was already in the security-idp-server module it was not showing up on the login page as the output target folder was being overwritten or overided. I fixed this by unpacking the ddf-theme in the security-idp-server module into it's own folder and then configuring the server to include that path at the /idp/ endpoint.