DDF-3621 Adapted EmbeddedSolrClientFactory to new factory implementation.
Updated to allow better customization of the embedded Solr server as required by the FilteringSolrIndex and ReuterSolrImport classes.
Reduced the complexity of the getEmbeddedSolrServer() method and fixed an issue where resources would not be closed if an error occurred while creating the embedded Solr server.
Modified ConfigurationStore to allow it to be created by the FilteringSolrIndex to avoid a bug where the changes it does to the singleton would have affected all Solr clients created after.