DDF-3844 Fixes Product overwrite versioning bug
There was a possibility previously that we would version the wrong metacard
when versioning a product overwrite. However it did not appear to impact us
because when quering for the already updated metacard we got returned a
cached version that hid the issue. Now we are using the already included
metacard (in the updateRequest) which we should have been doing all along.

Added missing mocks for HistorianTest