DDF-3855: Add CSRF Protections to Jolokia and Intrigue Endpoints
Adds the header "X-Requested-With" as a required header on the backend
so that all incoming requests to anything under /admin/jolokia
or /search/catalog/internal must contain this header or a 403 will be returned.
Value of this header can be anything, only it's presence it checked.
The incoming request must also have either an origin or referer header
that matches the target origin or a 403 will be returned.

Also adds this header to any necessary UI packages.