DDF-3621 Solr client/factory initial abstraction with common retry mechanism - final rebase (#3291)
Introduced new SolrClient interface and updated SolrClientFactory interface
Removed the Solr client availability management from the SolrCatalogProvider to take advantage of the new SolrClient interface.
Fixed dynamic schema resolver to ensure that fields were properly added once client is available as it used to fail silently if the server was down during the query and would never retry once it comes back.
Removed the Solr client availability management from the SolrCache to take advantage of the new SolrClient interface.
Fixed the initialization issue which was blocking bluprint until the client was available.
Updated PersistentStoreImpl to account for the fact that the new Solr factory will always returned a Solr client.
Adapted HttpSolrClientFactory to new factory implementation. Removed the retry mechanism as it is now handled by the SolrClientAdapter.
Prevented the low-level retry handler from retrying forever and made sure that ping requests would be retried at a low-level
Adapted EmbeddedSolrClientFactory to new factory implementation.
Updated to allow better customization of the embedded Solr server as required by the FilteringSolrIndex and ReuterSolrImport classes.
Reduced the complexity of the getEmbeddedSolrServer() method and fixed an issue where resources would not be closed if an error occurred while creating the embedded Solr server.
Modified ConfigurationStore to allow it to be created by the FilteringSolrIndex to avoid a bug where the changes it does to the singleton would have affected all Solr clients created after.
Updated SchemaFieldResolver and SolrQueryFilterVisitor based on new SolrClient API
Removed Banana wich is no longer supported. Updated timing on SolrProviderTest to account for slow systems initializing the Solr client