DDF-3855: CSRF Filter External Hostname Support
CSRF filter now compares the origin or referer header for
incoming requests to the hostname and port specified in
system.properties for local and external authorities. This allows
the CSRF to operate properly when behind a proxy that does not
modify origin or referer headers so long as the system.properties
is correctly configured to include the proxy's hostname + port.