DDF-3962 Resolve IE10 compatibility issue
The code the `routes-loader` generates doesn't get passed through babel,
so we cannot use ES6 features. We could figure out a way to pass the
code through babel, but it is sufficient to simply not use the features
for now.

The fetch api is unavailable in IE10, we can address this issue by using
a polyfill.

(cherry picked from commit bd03fc9)