DDF-4135 Don't overwite internal ports with external ports (#3728)
This commit changes the SystemPropertiesAdmin to no longer overwrite the
system ports whenever the external ports are changed. Now if the user
changes their external ports before install, the installer won't
overwrite the system ports. Changing the ports in the installer will
also only change the external ports, as it says it will.