DDF-4285 Validate certs before saving system properties
The installer used to save the system properties entered into it, then
check if the entered certs were valid. This could result in ddf running
on a different hostname than what was in the certs and make the
installer inaccessable. This commit will have the installer validate and save the certs
first, then save the system properties.