DDF-4357 Move state selection or clearing logic from `render` to `componentDidUpdate`
Modifying state during render is dangerous in React; `componentDidUpdate` can safely contain state modifications. Note that `componentDidUpdate` is not called upon first render, but the `Menu` constructor currently contains logic to select the first item in the menu if there are any. This first item selection logic in the constructor did not accomplish the goal of selecting the first item in the Gazetteer search box because that search box actually renders first with no children while the suggestions are being retrieved from Solr, then rerenders with new children without selecting the first of the new children.