DDF-04273 Implement DoSFilter
Adds the DelegateServletFilter back into DDF which will hold a DoSFilter
so that we can configure it. This will prevent jetty from overwriting our
configuration when it calls init on each of the filters injected in
a servlet since the DelegateServletFilters init method only takes the
servlet context from FilterConfig given to it. Also adds a
ProxyFilterChain that allows us to add Filters to a
javax.servlet.FilterChain, similar to how the ProxySecurityFilterChain
allows us to add SecurityFilters to a ddf.platform.filter.FilterChain.