DDF-04275 Replace keyczar with tink (#4431)
Keyczar uses SHA1, which is insecure. Keyczar should be removed and replaced with google's newer crypto library, Tink.
This PR takes out Keyczar, replaces it with Tink, and abstracts the underlying cryptography logic into a Crypter class. The Tink packages, and Tink's necessary transient dependencies's packages, are added as extensions to bundle 0 (except for in platform-osgi-configadmin where the dependencies are embedded because of weird package resolution issues). The Crypter jar is embedded where it is needed.