DDF-4716 Fix WFS 1.1.0 source's metatype.xml
The WFS 1.1.0 source allows its certificate alias, keystore, and SSL
protocol to be configured programmatically but not through the Admin UI
because its metatype.xml is missing them. This commit adds those three
properties to the metatype.xml so they can be configured in the Admin
UI. It also simplifies the refresh logic in the WFS 1.1.0 source so it
will reconnect and reconfigure the available WFS features every time
instead of having extensive logic to determine whether a reconnect is
necessary. Refreshes do not happen frequently and reconnecting and
reconfiguring the available WFS features is not that expensive.