DDF-4642 Update FeatureMetacardType to handle simpleTypes and simpleContent to arbitrary depths (#4643)
Previously, when evaluating a feature property whose type is a
simpleType, FeatureMetacardType would only go one level deep to
determine its ‘base’ type. The ‘base’ type is used to determine what
AttributeType to use to map to the feature property, e.g. string, date,
or integer. If the ‘base’ type is a type we can map to an AttributeType
then FeatureMetacardType adds the feature property to itself as an
attribute descriptor. Only going one level deep meant that feature
properties whose type was a ‘base’ schema type such as xsd:string or
xsd:int would be mapped and added as attribute descriptor, but feature
properties whose type was a custom-defined simpleType extending a ‘base’
type would not be because we’d look at its QName, see it’s not a ‘base’
type, and pass over it. As a result, feature properties whose type was a
simpleType were not queryable at all.

This commit changes how FeatureMetacardType handles simpleTypes. Now it
will attempt to find a simpleType’s root type by following <list> and
<restriction> definitions until a ‘base’ schema type is reached.

This commit also updates FeatureMetacardType to handle simpleContent
inside complexTypes. Now simpleContent <restriction>s and <extension>s
are handled in virtually the same way simpleTypes are handled as
described above.

The upshot is that WFS sources can now query feature properties whose
type is a simpleType or a complexType with simpleContent and that type’s
root type is a ‘base’ schema type that DDF can understand.

The three FeatureMetacardTypeTest classes in their respective WFS 1.0.0,
1.1.0, and 2.0.0 modules have been consolidated into one
FeatureMetacardTypeTest that lives in the spatial-wfs-common module.
FeatureMetacardType lives in the spatial-wfs-common module so the unit
tests for it in the other three modules did not count for unit test
coverage against FeatureMetacardType. Moreover, the tests in the three
classes were virtually identical so consolidating them into one made
sense.

Finally, the XmlSchema object used in the FeatureMetacardType unit tests
is now obtained by reading a real schema file instead of being
constructed manually. The schema object trees constructed in the tests
did not accurately model the schema object trees seen at runtime when a
real schema file is parsed. Now that the unit tests obtain the XmlSchema
object in the same way the runtime code does, they will be identical.