DDF-4659 Make HandlebarsWfsFeatureTransformer handle XML attributes (#4664)
HandlebarsWfsFeatureTransformer now adds the XML attributes of all the
elements in the feature to its internal map. The key is of the form
elementName@attributeName to make sure attributes with the same name on
different elements do not overwrite each other. In your mapping JSON you
can specify "elementName@attributeName" as your feature name and use it
in your Handlebars template to get the attribute's value.