DDF-4983 Fix check for Solr RealTime Get (#5045)
This refines the check to see if we should do a Solr RealTime Get so that we
only do it for filters that only contain checks for metacards with
specific ids. Before, we did it for any filter that contained an "id ="
operand, which could cause a query to only return metacards matching an
id when the filter should return other results as well.