DDF-5305 Support query mappability in WFS 1.0 source
The WFS 1.0.0 source now supports the use of an optional MetacardMapper
service to convert metacard attributes to WFS feature properties for
queries to the WFS server. It also supports the use of the
MetacardMapper service to convert WFS feature properties to metacard
attributes when transforming query responses from the WFS server.

This commit also fixes the WFS 1.0.0 connected source bundle so it is
now usable again.