DDF-5132 Replace logout UI component with single logout page
Previously the UI used a JS component which handled logging the user
out. The "sign out" buttons throughout the UI called out to it to
handle logouts. This worked for frontend code, but there was no
way for the backend to use it.

This commit modifies the logout page to perform the same functions
as the JS component. Now, both the frontend and backend can use it
by redirecting the user to this page.

In particular, this fixes the session management service's
invalidate() method.