DDF-4750 Fix WFS 1.1.0 source paging (#5365)
When a WFS 1.1.0 source is queried, it will now send a 'hits' query to
the WFS server before it requests the results. (In WFS 1.1.0 there is no
way to request the total hits and some number of results at the same
time.) The source will now return the actual hit total in the response.
Paging will now behave as expected on a WFS 1.1.0 source since the
source knows the actual hit total.