DDF-5432 Fix point-radius lat/lon input validation
Currently you cannot enter a latitude or longitude that starts with 0
(positive or negative) or even 0 itself. point-radius.js sets the
field's value to '' if the actual input value is 0 so as soon as you
enter 0 or a value that resolves to 0 (like -0), it is erased. That
solution was implemented to fix a bug that caused the fields to display
0 whenever the input was fully erased. The root cause is that the
Math.max()/Math.min() operations performed on the input value to fit it
in the valid latitude or longitude range return 0 for '' (the value of
a fully-erased input field) which results in the field being set to 0
even if you erase the whole thing. This solution simply adds a check
for the empty string before attempting to fit the value to the valid
range which removes the need to check for 0 in point-radius.js.