DDF-5455 Improve GML parsing performance (#5587)
Gml3ToWktImpl now uses a SAXParser with GeoTools' GMLParserDelegate
instead of GeoTools' Parser class. The Parser class creates a new
SAXParserFactory on every parse() call which results in bad performance
when batches of GML are parsed. Now, Gml3ToWktImpl creates a single
SAXParserFactory and uses it to create SAXParsers (which is expected to
be thread safe according to the JAXP 1.6 spec, section 4.14) which use
GeoTools' GMLParserDelegate to handle the GML, so Gml3ToWktImpl's
functionality should remain unchanged.