Enable longitude wrapping on the indexed Solr geo field
Running a DWITHIN query (i.e., a buffered geometry) near the
antimeridian often results in a Solr error because the resulting
geometry contains points outside the valid longitude range of [-180,180].
The SolrFilterDelegate uses JTS to calculate the buffered geometry but
does not enforce the longitude range on the result so it is passed to
Solr with out-of-range values. Enabling normWrapLongitude means that
Solr will wrap longitudes outside of the range into it.