Remove _root_ field from Solr schema (#6602) (#6605)
Solr 8.x will always populate the _root_ field with the unique key
(which is id_txt for DDF), if that field is in the schema. When Solr 8.x
detects the presence of the _root_ field in the schema, it will use that
field to query for Lucene documents when making updates. Solr 7.x and
earlier does not populate that field, so a Solr 7.x index will have the
field in its schema but not in its Lucene index (assuming the nested
documents feature is not being actively used, and DDF does not use it).
This causes Solr 8.x to think the document does not exist, so it creates
a new Lucene document with a different Lucene Doc ID but the same Solr
ID. SolrCloud appears to de-duplicate by ID, so only the old stored
document is returned when there is a match for either Lucene document.

This means that if you are upgrading a Solr 7.x or earlier to Solr 8.x,
any existing documents cannot be updated if _root_ is in the schema,
since a new document will be created for each on the first update, but
queries will return the original documents.

Downstream projects will need to add the _root_ field back to the schema
if they are using nested documents and are using DDF's schema.