Prevent AND queries containing invalid filters from running against WFS 1.1.0 sources
Now, the WfsFilterDelegate will return null when ANDing filters and at
least one is null. This means any AND group will be dropped if it
contains a feature property not understood by the source being queried.
If this results in no filter for the query at all, then it will not be
run against the source.

Also, note that test coverage did not actually decrease with this
change. The instruction coverage was updated to reflect its true value;
LenientLimit has prevented this from failing the build.